requirements.txt
Flask
Biopython
numpy
scipy
matplotlib
multidict
oauthlib
opt-einsum
pandas
protobuf
pyasn1
pyasn1-modules
pyparsing
scikit-learn
seaborn
tensorflow>=2
tensorboard-plugin-wit
termcolor
threadpoolctl
urllib3
wrapt
yarl
tensorflow-addons
mysql-connector-python